% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aebar.R
\name{aebar}
\alias{aebar}
\title{Bar chart for number of events reported per participant}
\usage{
aebar(
  data,
  arm_levels,
  id = "id",
  arm = "arm",
  arm_names = NULL,
  arm_colours = NULL,
  facets = TRUE,
  save_image_path = NULL
)
}
\arguments{
\item{data}{data frame with adverse_event, id and arm columns}

\item{arm_levels}{vector of factor levels in arm variable}

\item{id}{name of id column}

\item{arm}{name of arm column}

\item{arm_names}{vector of names for each arm in arm variable}

\item{arm_colours}{vector of colours for each arm}

\item{facets}{a logical value whether to plot bar chart with facets}

\item{save_image_path}{file path to save bar chart as image}
}
\value{
Bar chart presenting the number of events reported per participant
}
\description{
Bar chart for number of events reported per participant
}
\examples{
aebar(df2, arm_levels=c("Intervention","Placebo"), facets=FALSE)
}
