% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aedot.R
\name{aedot}
\alias{aedot}
\title{Dot plot to visualise AE and harm profiles in two-arm randomised controlled trials}
\usage{
aedot(
  data,
  body_system_class = "body_system_class",
  id = "id",
  arm = "arm",
  arm1 = "A1",
  arm2 = "A2"
)
}
\arguments{
\item{data}{Dataframe with adverse_event, body_system_class, id and arm columns}

\item{body_system_class}{Name of body_system_class column}

\item{id}{Name of id column}

\item{arm}{Name of arm column}

\item{arm1}{Factor level of arm 1}

\item{arm2}{Factor level of arm 2}
}
\value{
Dot plot with proportions alongside treatment effect estimates (IRR) with accompanying 95\% confidence interval
}
\description{
Dot plot to visualise AE and harm profiles in two-arm randomised controlled trials
}
\examples{
aedot(df, body_system_class="ae_02", id="participant_id", arm="treatment_arm")
}
