% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemcf.R
\name{aemcf}
\alias{aemcf}
\title{Mean cumulative function plot for harm outcomes by treatment group}
\usage{
aemcf(
  data,
  arm_levels,
  subset,
  adverse_event = "adverse_event",
  body_system_class = "body_system_class",
  id = "id",
  arm = "arm",
  date_rand = "date_rand",
  date_ae = "date_ae",
  last_visit = "last_visit",
  time_units = "days",
  arm_names = NULL,
  arm_colours = NULL,
  conf.int = TRUE,
  risk_table = TRUE,
  save_image_path = NULL
)
}
\arguments{
\item{data}{data frame with adverse_event, body_system_class, id, arm, date_rand, date_ae and last_visit columns}

\item{arm_levels}{vector of factor levels in arm variable}

\item{subset}{an optional argument specifying a subset of observations to be used}

\item{adverse_event}{name of adverse_event column}

\item{body_system_class}{name of body_system_class column}

\item{id}{name of id column}

\item{arm}{name of arm column}

\item{date_rand}{name of date_rand column}

\item{date_ae}{name of date_ae column}

\item{last_visit}{name of last_visit column}

\item{arm_names}{vector of names for each arm in arm variable}

\item{arm_colours}{vector of colours for each arm}

\item{conf.int}{a logical value whether to plot the 95\% confidence intervals}

\item{risk_table}{a logical value whether to plot the risk table}

\item{save_image_path}{file path to save MCF plot as image}
}
\value{
MCF plot with risk table
}
\description{
\code{aemcf} plots the mean cumulative function (MCF) of all adverse events, adverse events of specific organ systems or specific adverse events with embedded table of number of participants at risk.
The nonparametric MCF estimates (by Nelson-Aalen estimator) and the approximate 95\% confidence intervals are obtained using the \code{mcf} function of the \code{reda} package.
}
\examples{
aemcf(df2, c("Placebo", "Intervention"), subset= body_system_class=="Gastrointestinal", adverse_event="ae_pt", body_system_class="aebodsys")
}
