% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeseverity.R
\name{aeseverity}
\alias{aeseverity}
\title{Table of frequencies and proportions of events by severity categories}
\usage{
aeseverity(
  data,
  severity = "severity",
  id = "id",
  arm = "arm",
  arm1 = "A1",
  arm2 = "A2",
  arm3 = "A3",
  arm4 = "A4",
  arm1_name = "Arm 1",
  arm2_name = "Arm 2",
  arm3_name = "Arm 3",
  arm4_name = "Arm 4",
  save_image_path = NULL,
  save_docx_path = NULL
)
}
\arguments{
\item{data}{Dataframe with adverse event, severity, id and arm columns}

\item{severity}{Name of severity column}

\item{id}{Name of id column}

\item{arm}{Name of arm column}

\item{arm1}{Factor level of arm 1}

\item{arm2}{Factor level of arm 2}

\item{arm3}{Factor level of arm 3}

\item{arm4}{Factor level of arm 4}

\item{arm1_name}{Name of arm 1}

\item{arm2_name}{Name of arm 2}

\item{arm3_name}{Name of arm 3}

\item{arm4_name}{Name of arm 4}

\item{save_image_path}{file path to save table as image}

\item{save_docx_path}{file path to save table as docx}
}
\value{
Flextable of frequencies and proportions of events by severity categories
}
\description{
Table of frequencies and proportions of events by severity categories
}
