% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aestacked.R
\name{aestacked}
\alias{aestacked}
\title{Stacked bar chart of proportions for each body system class by arm and maximum severity}
\usage{
aestacked(
  data,
  body_system_class = "body_system_class",
  severity = "severity",
  id = "id",
  arm = "arm",
  arm1 = "A1",
  arm2 = "A2",
  arm3 = "A3",
  arm4 = "A4",
  arm1_name = "Arm 1",
  arm2_name = "Arm 2",
  arm3_name = "Arm 3",
  arm4_name = "Arm 4",
  severity_levels = c(),
  severity_colours = c(),
  save_image_path = NULL
)
}
\arguments{
\item{data}{data frame with adverse_event, body_system_class, severity, id and arm columns}

\item{body_system_class}{name of body_system_class column}

\item{severity}{name of severity column}

\item{id}{name of id column}

\item{arm}{name of arm column}

\item{arm1}{factor level of arm 1}

\item{arm2}{factor level of arm 2}

\item{arm3}{factor level of arm 3}

\item{arm4}{factor level of arm 4}

\item{arm1_name}{name of arm 1}

\item{arm2_name}{name of arm 2}

\item{arm3_name}{name of arm 3}

\item{arm4_name}{name of arm 4}

\item{severity_levels}{vector of level of severity in ascending order if severity is not ordered factor}

\item{severity_colours}{vector of colours for level of severity in ascending order}

\item{save_image_path}{file path to save stacked bar chart plot as image}
}
\value{
stacked bar chart presenting the proportions of participants with each event by arm and by maximum severity
}
\description{
Stacked bar chart of proportions for each body system class by arm and maximum severity
}
\examples{
aestacked(df, body_system_class="ae_02", severity="ae_05", arm1="Anti-IgE", arm2="Placebo", severity_levels=c("Mild", "Moderate", "Severe"), arm1_name="Anti-IgE", arm2_name="Placebo")
}
