% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aetable.R
\name{aetable}
\alias{aetable}
\title{Table of AE summary by body system class}
\usage{
aetable(
  data,
  body_system_class = "body_system_class",
  id = "id",
  arm = "arm",
  date_rand = "date_rand",
  last_visit = "last_visit"
)
}
\arguments{
\item{data}{Dataframe with adverse_event, body_system_class, id, arm, date_rand and last_visit columns}

\item{body_system_class}{Name of body_system_class column}

\item{id}{Name of id column}

\item{arm}{Name of arm column}

\item{date_rand}{Name of date_rand column}

\item{last_visit}{Name of last_visit column}
}
\value{
Flextable of AE summary by body system class
}
\description{
Table of AE summary by body system class
}
