% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df3}
\alias{df3}
\title{Sample adverse event dataset with 3 treatment arms}
\format{
A data frame with 370 rows and 6 variables:
\describe{
\item{ae_01}{Adverse event}
\item{ae_02}{Body system class}
\item{id}{Participant's id}
\item{arm}{Treatment arm}
\item{date_rand}{Date of randomisation}
\item{last_visit}{Date of last visit}
\item{ae_05}{Severity}
}
}
\usage{
df3
}
\description{
Sample adverse event dataset with 3 treatment arms
}
\keyword{datasets}
