% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df4}
\alias{df4}
\title{Sample adverse event dataset with 4 treatment arms}
\format{
A data frame with 370 rows and 8 variables:
\describe{
\item{id}{Participant's id}
\item{arm}{Treatment arm: Placebo, Intervention 1, Intervention 2, Intervention 3}
\item{ae_pt}{Adverse event preferred term}
\item{aebodsys}{Body system class}
\item{severity}{Severity}
\item{date_rand}{Date of randomisation}
\item{date_ae}{Date of adverse event}
\item{last_visit}{Date of last visit}
}
}
\usage{
df4
}
\description{
Sample adverse event dataset with 4 treatment arms
}
\keyword{datasets}
