% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lab2}
\alias{lab2}
\title{Sample laboratory values with continuous outcomes dataset with 2 treatment arms}
\format{
A data frame with 2580 rows and 9 variables:
\describe{
\item{id}{Participant's id}
\item{arm}{Treatment arm: Placebo & Intervention}
\item{visit}{Timepoint of measurement}
\item{lab_test}{Laboratory test}
\item{base}{Baseline value of measurement}
\item{aval}{Measurement value}
\item{region}{Participant's region}
\item{strat}{Stratification variable}
\item{Time}{Time of measurement}
}
}
\usage{
lab2
}
\description{
Sample laboratory values with continuous outcomes dataset with 2 treatment arms
}
\keyword{datasets}
