% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labscatter.R
\name{labscatter}
\alias{labscatter}
\title{Scatterplot matrix for baseline vs follow-up laboratory values by treatment group}
\usage{
labscatter(
  data,
  arm_levels,
  lab_test_list,
  limit_list,
  cutoff_list,
  lab_test = "lab_test",
  visit = "visit",
  id = "id",
  arm = "arm",
  base = "base",
  aval = "aval",
  lower,
  upper,
  last_visit = FALSE,
  arm_names = NULL,
  arm_colours = NULL,
  save_image_path = NULL
)
}
\arguments{
\item{data}{dataframe with lab_test, visit, id, arm, base and aval columns (optional: lower and upper limit columns)}

\item{arm_levels}{vector of factor levels in arm variable}

\item{lab_test_list}{a character vector of laboratory test names to be included in the scatterplot matrix}

\item{limit_list}{a character vector specifying the direction of the cutoff for each laboratory test in \code{lab_test_list}; each element should be either \code{"upper"} or \code{"lower"}}

\item{cutoff_list}{a numeric vector of cutoff values (upper or lower) for each laboratory test in \code{lab_test_list}}

\item{lab_test}{name of lab_test column}

\item{visit}{name of visit column}

\item{id}{name of id column}

\item{arm}{name of arm column}

\item{base}{name of base column}

\item{aval}{name of aval column}

\item{lower}{name of lower column}

\item{upper}{name of upper column}

\item{last_visit}{a logical value whether to plot against final visit laboratory measurement value instead of maximum/minimum across all follow-up visits}

\item{arm_names}{a character vector of names for each arm in arm variable}

\item{arm_colours}{a vector of colours for each arm}

\item{save_image_path}{file path to save scatterplot matrix as image}
}
\value{
scatterplot matrix of
}
\description{
\code{labscatter} generates a scatterplot matrix to visualise multiple continuous harm outcomes by treatment group.
For each specified laboratory test, the function plots individual participants' baseline values against their maximum or minimum post-baseline values, or optionally their final visit value.
Dashed lines represent the boundary between normal and abnormal thresholds, which can be defined using the \code{cutoff_list} argument or drawn from the the \code{lower} and \code{upper} columns in the dataset.
Outliers beyond these thresholds are labelled with participant IDs.
}
\examples{
labscatter(lab2, arm_levels=c("Placebo", "Intervention"), lab_test_list=c("Lymphocytes (GI/L)", "Monocytes (GI/L)", "Potassium (mmol/L)"), limit_list=c("upper", "upper", "lower"), cutoff_list=c(4.1, 1.1, 3.5))
}
