% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aedot.R
\name{aedot}
\alias{aedot}
\title{Dot plot to visualise AE and harm profiles in two-arm randomised controlled trials}
\usage{
aedot(
  data,
  body_system_class = "body_system_class",
  id = "id",
  arm = "arm",
  control = "C",
  intervention = "I",
  save_image_path = NULL
)
}
\arguments{
\item{data}{data frame with adverse_event, body_system_class, id and arm columns}

\item{body_system_class}{name of body_system_class column}

\item{id}{name of id column}

\item{arm}{name of arm column}

\item{control}{factor level of control arm}

\item{intervention}{factor level of intervention arm}

\item{save_image_path}{file path to save dot plot as image}
}
\value{
Dot plot with proportions alongside treatment effect estimates (IRR) with accompanying 95\% confidence interval
}
\description{
Dot plot to visualise AE and harm profiles in two-arm randomised controlled trials
}
\examples{
aedot(df, body_system_class="ae_02", control="Placebo", intervention="Anti-IgE")
}
