% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aedot.R
\name{aedot}
\alias{aedot}
\title{Dot plot to visualise AE and harm profiles in two-arm randomised controlled trials}
\usage{
aedot(
  data,
  control,
  intervention,
  body_system_class = "body_system_class",
  id = "id",
  arm = "arm",
  date_rand = "date_rand",
  last_visit = "last_visit",
  model = "unadjusted (RR)",
  variables = c(),
  dot_colours = c("red", "blue"),
  save_image_path = NULL
)
}
\arguments{
\item{data}{data frame with adverse_event, body_system_class, id and arm columns (optional: date_rand, last_visit and variables to be included in model)}

\item{control}{factor level of control arm}

\item{intervention}{factor level of intervention arm}

\item{body_system_class}{name of body_system_class column}

\item{id}{name of id column}

\item{arm}{name of arm column}

\item{date_rand}{name of date_rand column}

\item{last_visit}{name of last_visit column}

\item{model}{unadjusted relative risk/risk difference or model used for computation of treatment effect estimate and 95\% CI}

\item{variables}{vector of variable names to be included in the model for computation of treatment effect estimate (excluding arm)}

\item{dot_colours}{vector of colours for dots on percentage of participants plot (left plot)}

\item{save_image_path}{file path to save dot plot as image}
}
\value{
dot plot with proportions alongside treatment effect estimates with accompanying 95\% confidence interval
}
\description{
\code{aedot} creates a plot that summarises the AE and harm profiles via the following visualisations:
\itemize{
\item \strong{left panel}: percentage of participants who had an event in each treatment group
\item \strong{central panel}: relative risk/ risk difference/ incidence rate ratio/ odds ratio of observing each event in the treatment group compared with the control group, along with 95\% confidence interval
\item \strong{right panel}: data table containing number of participants with at least one event and number of events by treatment group
}
}
\section{Treatment Effect Estimate}{

The treatment effect and its 95\% confidence interval are estimated via the following models below:
\itemize{
\item \strong{unadjusted (RR)}: unadjusted relative risk
\item \strong{unadjusted (RD)}: unadjusted risk difference
\item \strong{Poisson (rate)}: fitting a generalised linear model with Poisson family and log link with length of follow up time as offset
\item \strong{Poisson (count)}: fitting a generalised linear model with Poisson family and log link with count as response
\item \strong{Negative Binomial (rate)}: fitting a Negative Binomial model with length of follow up time as offset
\item \strong{Negative Binomial (count)}: fitting a Negative Binomial model with count as response
\item \strong{Binomial (logit)}: fitting a generalised linear model with Binomial family and logit link
\item \strong{Binomial (log)}: fitting a generalised linear model with Binomial family and log link
\item \strong{Binomial (identity)}: fitting a generalised linear model with Binomial family and identity link
}

The default model is the \code{unadjusted (RR)} model.

For Poisson and Negative Binomial models, the treatment effect estimate is the \strong{Incidence Rate Ratio (IRR)}.
For Binomial (logit) model, the treatment effect estimate is the \strong{Odds Ratio (OR)}.
For Binomial (log) model, the treatment effect estimate is the \strong{Relative Risk (RR)}.
For Binomial (identity) model, the treatment effect estimate is the \strong{Risk Difference (RD)}.

The reference level for the arm variable used in the model is the control arm.

Additional covariates besides arm can be added into the model via the argument \code{variables}.
Note that interaction terms cannot be added to the model.
}

\examples{
df2$aebodsys <- as.factor(df2$aebodsys)
aedot(df2, body_system_class="aebodsys", control="Placebo", intervention="Intervention")
}
