% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeseverity.R
\name{aeseverity}
\alias{aeseverity}
\title{Table of frequencies and proportions of events by severity categories}
\usage{
aeseverity(
  data,
  arm_levels,
  severity_levels = NULL,
  body_system_class = "body_system_class",
  severity = "severity",
  id = "id",
  arm = "arm",
  n_events = TRUE,
  arm_names = NULL,
  severity_names = NULL,
  proportions_dp = 1,
  save_image_path = NULL,
  save_docx_path = NULL
)
}
\arguments{
\item{data}{data frame with adverse_event, body_system_class, severity, id and arm columns}

\item{arm_levels}{vector of factor levels in arm variable}

\item{severity_levels}{vector of severity levels in ascending order if severity is not ordered factor}

\item{body_system_class}{name of body_system_class column}

\item{severity}{name of severity column}

\item{id}{name of id column}

\item{arm}{name of arm column}

\item{n_events}{a logical value whether to include number of events for each body system class and severity}

\item{arm_names}{vector of names for each arm in arm variable}

\item{severity_names}{vector of names for each severity in severity variable}

\item{proportions_dp}{number of decimal places for proportions}

\item{save_image_path}{file path to save table as image}

\item{save_docx_path}{file path to save table as docx}
}
\value{
flextable of frequencies and proportions of events by severity categories
}
\description{
\code{aeseverity} is used to plot a table of AE summary by severity level and body system class for each arm using the following summary statistics:
\itemize{
\item \strong{N}: Frequency (number of participants with at least one event for each severity level and body system class)
\item \strong{\%}: Proportions (number of participants with at least one event for each severity level and body system class relative to number of participants at risk)
\item \strong{n}: Number of adverse events for each severity level and body system class
}

Note that \code{aeseverity} can only take up to 5 severity levels for the severity variable.
}
\examples{
library(flextable)
set_flextable_defaults(font.family="sans")
df2$aebodsys <- as.factor(df2$aebodsys)
df2$severity <- ordered(df2$severity, c("Mild", "Moderate", "Severe"))
aeseverity(df2, arm_levels=c("Intervention","Placebo"), body_system_class="aebodsys", proportions_dp=2)
}
