% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeseverity.R
\name{aeseverity}
\alias{aeseverity}
\title{Table of frequencies and proportions of events by severity categories}
\usage{
aeseverity(
  data,
  severity = "severity",
  id = "id",
  arm = "arm",
  arm_levels = c("A1", "A2", "A3", "A4"),
  arm_names = NULL,
  proportions_dp = 1,
  save_image_path = NULL,
  save_docx_path = NULL
)
}
\arguments{
\item{data}{data frame with adverse event, severity, id and arm columns}

\item{severity}{name of severity column}

\item{id}{name of id column}

\item{arm}{name of arm column}

\item{arm_levels}{vector of factor levels in arm variable}

\item{arm_names}{vector of names for each arm in arm variable}

\item{proportions_dp}{number of decimal places for proportions}

\item{save_image_path}{file path to save table as image}

\item{save_docx_path}{file path to save table as docx}
}
\value{
flextable of frequencies and proportions of events by severity categories
}
\description{
Table of frequencies and proportions of events by severity categories
}
\examples{
df2$severity <- ordered(df2$severity, c("Mild", "Moderate", "Severe"))
aeseverity(df2, arm_levels=c("Intervention","Placebo"), proportions_dp=2)
}
