% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aestacked.R
\name{aestacked}
\alias{aestacked}
\title{Stacked bar chart of proportions for each body system class by arm and maximum severity}
\usage{
aestacked(
  data,
  arm_levels,
  body_system_class = "body_system_class",
  severity = "severity",
  id = "id",
  arm = "arm",
  arm_names = NULL,
  severity_levels = NULL,
  severity_colours = NULL,
  save_image_path = NULL
)
}
\arguments{
\item{data}{data frame with adverse_event, body_system_class, severity, id and arm columns}

\item{arm_levels}{vector of factor levels in arm variable}

\item{body_system_class}{name of body_system_class column}

\item{severity}{name of severity column}

\item{id}{name of id column}

\item{arm}{name of arm column}

\item{arm_names}{vector of names for each arm in arm variable}

\item{severity_levels}{vector of level of severity in ascending order if severity is not ordered factor}

\item{severity_colours}{vector of colours for level of severity in ascending order}

\item{save_image_path}{file path to save stacked bar chart plot as image}
}
\value{
stacked bar chart presenting the proportions of participants with each event by arm and by maximum severity
}
\description{
Stacked bar chart of proportions for each body system class by arm and maximum severity
}
\examples{
df2$aebodsys <- as.factor(df2$aebodsys)
df2$severity <- ordered(df2$severity, c("Mild", "Moderate", "Severe"))
aestacked(df2, body_system_class="aebodsys", arm_levels=c("Intervention","Placebo"), severity_levels=c("Mild", "Moderate", "Severe"))
}
