% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aetable.R
\name{aetable}
\alias{aetable}
\title{Table of AE summary by body system class}
\usage{
aetable(
  data,
  body_system_class = "body_system_class",
  id = "id",
  arm = "arm",
  date_rand = "date_rand",
  last_visit = "last_visit",
  arm1 = "A1",
  arm2 = "A2",
  arm3 = "A3",
  arm4 = "A4",
  arm1_name = "Arm 1",
  arm2_name = "Arm 2",
  arm3_name = "Arm 3",
  arm4_name = "Arm 4",
  save_image_path = NULL,
  save_docx_path = NULL
)
}
\arguments{
\item{data}{Dataframe with adverse_event, body_system_class, id, arm, date_rand and last_visit columns}

\item{body_system_class}{Name of body_system_class column}

\item{id}{Name of id column}

\item{arm}{Name of arm column}

\item{date_rand}{Name of date_rand column}

\item{last_visit}{Name of last_visit column}

\item{arm1}{Factor level of arm 1}

\item{arm2}{Factor level of arm 2}

\item{arm3}{Factor level of arm 3}

\item{arm4}{Factor level of arm 4}

\item{arm1_name}{Name of arm 1}

\item{arm2_name}{Name of arm 2}

\item{arm3_name}{Name of arm 3}

\item{arm4_name}{Name of arm 4}

\item{save_image_path}{file path to save table as image}

\item{save_docx_path}{file path to save table as docx}
}
\value{
Flextable of AE summary by body system class
}
\description{
Table of AE summary by body system class
}
\examples{
aetable(df, body_system_class="ae_02", id="participant_id", arm="treatment_arm", date_rand="randomisation_date", last_visit="date_of_last_visit")
}
