% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aetable.R
\name{aetable}
\alias{aetable}
\title{Table of AE summary by body system class}
\usage{
aetable(
  data,
  body_system_class = "body_system_class",
  id = "id",
  arm = "arm",
  date_rand = "date_rand",
  last_visit = "last_visit",
  control = "C",
  intervention1 = "I1",
  intervention2 = "I2",
  intervention3 = "I3",
  control_name = "Control",
  intervention1_name = "Intervention 1",
  intervention2_name = "Intervention 2",
  intervention3_name = "Intervention 3",
  IRR = TRUE,
  variables = c(),
  mean = TRUE,
  proportions_dp = 1,
  IR_dp = 1,
  mean_dp = 1,
  SD_dp = 1,
  IRR_sf = 3,
  CI_sf = 3,
  save_image_path = NULL,
  save_docx_path = NULL
)
}
\arguments{
\item{data}{Dataframe with adverse_event, body_system_class, id, arm, date_rand and last_visit columns}

\item{body_system_class}{Name of body_system_class column}

\item{id}{Name of id column}

\item{arm}{Name of arm column}

\item{date_rand}{Name of date_rand column}

\item{last_visit}{Name of last_visit column}

\item{control}{Factor level of control arm}

\item{intervention1}{Factor level of intervention 1 arm}

\item{intervention2}{Factor level of intervention 2 arm}

\item{intervention3}{Factor level of intervention 3 arm}

\item{control_name}{Name of control arm}

\item{intervention1_name}{Name of intervention 1 arm}

\item{intervention2_name}{Name of intervention 2 arm}

\item{intervention3_name}{Name of intervention 3 arm}

\item{IRR}{Boolean to include IRR and 95\% CI column in summary table (only for 2 arms)}

\item{variables}{Vector of variables to be included in the glm Poisson model for computation of IRR besides arm}

\item{mean}{Boolean to include mean and SD column in summary table}

\item{proportions_dp}{Number of decimal places for proportions}

\item{IR_dp}{Number of decimal places for incidence rate}

\item{mean_dp}{Number of decimal places for mean number of AEs per participant}

\item{SD_dp}{Number of decimal places for standard deviation of number of AEs per participant}

\item{IRR_sf}{Number of significant figures for IRR}

\item{CI_sf}{Number of significant figures for 95\% CI}

\item{save_image_path}{File path to save table as image}

\item{save_docx_path}{File path to save table as docx}
}
\value{
Flextable of AE summary by body system class
}
\description{
Table of AE summary by body system class
}
\examples{
aetable(df, body_system_class="ae_02", id="participant_id", arm="treatment_arm", date_rand="randomisation_date", last_visit="date_of_last_visit")
}
