% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aetable.R
\name{aetable}
\alias{aetable}
\title{Table of AE summary by body system class}
\usage{
aetable(
  data,
  control,
  intervention_levels,
  body_system_class = "body_system_class",
  id = "id",
  arm = "arm",
  date_rand = "date_rand",
  last_visit = "last_visit",
  control_name = NULL,
  intervention_names = NULL,
  treatment_effect_estimate = TRUE,
  model = "Poisson (rate)",
  variables = c(),
  mean = TRUE,
  drop_bodsys = TRUE,
  IR_per_person = 100,
  proportions_dp = 1,
  IR_dp = 1,
  mean_dp = 1,
  SD_dp = 1,
  estimate_sf = 3,
  CI_sf = 3,
  save_image_path = NULL,
  save_docx_path = NULL
)
}
\arguments{
\item{data}{data frame with adverse_event, body_system_class, id, arm, date_rand and last_visit columns (optional: variables to be included in model)}

\item{control}{factor level of control arm}

\item{intervention_levels}{vector of factor levels for interventions}

\item{body_system_class}{name of body_system_class column}

\item{id}{name of id column}

\item{arm}{name of arm column}

\item{date_rand}{name of date_rand column}

\item{last_visit}{name of last_visit column}

\item{control_name}{name of control arm}

\item{intervention_names}{vector of names for interventions}

\item{treatment_effect_estimate}{a logical value whether to include treatment effect estimate and 95\% CI column in summary table (only for 2 arms)}

\item{model}{model used for computation of treatment effect estimate and 95\% CI}

\item{variables}{vector of variable names to be included in the model for computation of treatment effect estimate (excluding arm)}

\item{mean}{a logical value whether to include mean and SD column in summary table}

\item{drop_bodsys}{a logical value whether to drop body system class with no observations from summary table}

\item{IR_per_person}{incidence rate per number of person}

\item{proportions_dp}{number of decimal places for proportions}

\item{IR_dp}{number of decimal places for incidence rate}

\item{mean_dp}{number of decimal places for mean number of AEs per participant}

\item{SD_dp}{number of decimal places for standard deviation of number of AEs per participant}

\item{estimate_sf}{number of significant figures for treatment effect estimate}

\item{CI_sf}{number of significant figures for 95\% CI}

\item{save_image_path}{file path to save table as image}

\item{save_docx_path}{file path to save table as docx}
}
\value{
flextable of AE summary by body system class and arm
}
\description{
\code{aetable} is used to plot a table of AE summary by body system class and arm using the following summary statistics:
\itemize{
\item \strong{Total number of participants at risk per arm}
\item \strong{Frequency} (number of participants with at least one event for each body system class)
\item \strong{Proportions} (number of participants with at least one event for each body system class relative to number of participants at risk)
\item \strong{Number of adverse events per participant}: presented as \strong{counts}, \strong{mean (SD)}
\item \strong{Total number of events} and \strong{incidence rates} (\strong{number of events} relative to \strong{total time in follow-up})
\item \strong{Treatment effect estimate} and \strong{95\% confidence intervals}
}
}
\section{Treatment Effect Estimate}{

The treatment effect and its 95\% confidence interval are estimated via the following models below:
\itemize{
\item \strong{Poisson (rate)}: fitting a generalised linear model with Poisson family and log link with length of follow up time as offset
\item \strong{Poisson (count)}: fitting a generalised linear model with Poisson family and log link with count as response
\item \strong{Negative Binomial (rate)}: fitting a Negative Binomial model with length of follow up time as offset
\item \strong{Negative Binomial (count)}: fitting a Negative Binomial model with count as response
\item \strong{Binomial (logit)}: fitting a generalised linear model with Binomial family and logit link
\item \strong{Binomial (log)}: fitting a generalised linear model with Binomial family and log link
\item \strong{Binomial (identity)}: fitting a generalised linear model with Binomial family and identity link
}

The default model is the \code{Poisson (rate)} model.

For Poisson and Negative Binomial models, the treatment effect estimate is the \strong{Incidence Rate Ratio (IRR)}.
For Binomial (logit) model, the treatment effect estimate is the \strong{Odds Ratio (OR)}.
For Binomial (log) model, the treatment effect estimate is the \strong{Risk Ratio (RR)}.
For Binomial (identity) model, the treatment effect estimate is the \strong{Risk Difference (RD)}.

The reference group for the treatment arm in the regression model is the control arm.

Additional covariates besides arm can be added into the model via the argument \code{variables}.
Note that interaction terms cannot be added to the model.
}

\examples{
library(extrafont)
font_import(prompt=FALSE)
loadfonts(device="postscript", quiet=TRUE)
df2$aebodsys <- as.factor(df2$aebodsys)
aetable(df2, body_system_class="aebodsys", control="Placebo", intervention_levels=c("Intervention"), treatment_effect_estimate=TRUE, variables = c("variable1", "variable2"))
}
