% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aetable.R
\name{aetable}
\alias{aetable}
\title{Table of AE summary by body system class}
\usage{
aetable(
  data,
  body_system_class = "body_system_class",
  id = "id",
  arm = "arm",
  date_rand = "date_rand",
  last_visit = "last_visit",
  control = "C",
  intervention_levels = c("I1", "I2", "I3"),
  control_name = NULL,
  intervention_names = NULL,
  IRR = TRUE,
  variables = c(),
  mean = TRUE,
  proportions_dp = 1,
  IR_dp = 1,
  mean_dp = 1,
  SD_dp = 1,
  IRR_sf = 3,
  CI_sf = 3,
  save_image_path = NULL,
  save_docx_path = NULL
)
}
\arguments{
\item{data}{data frame with adverse_event, body_system_class, id, arm, date_rand and last_visit columns}

\item{body_system_class}{name of body_system_class column}

\item{id}{name of id column}

\item{arm}{name of arm column}

\item{date_rand}{name of date_rand column}

\item{last_visit}{name of last_visit column}

\item{control}{factor level of control arm}

\item{intervention_levels}{vector of factor levels for interventions}

\item{control_name}{name of control arm}

\item{intervention_names}{vector of names for interventions}

\item{IRR}{a logical value whether to include IRR and 95\% CI column in summary table (only for 2 arms)}

\item{variables}{vector of variable names to be included in the glm Poisson model for computation of IRR (excluding arm)}

\item{mean}{a logical value whether to include mean and SD column in summary table}

\item{proportions_dp}{number of decimal places for proportions}

\item{IR_dp}{number of decimal places for incidence rate}

\item{mean_dp}{number of decimal places for mean number of AEs per participant}

\item{SD_dp}{number of decimal places for standard deviation of number of AEs per participant}

\item{IRR_sf}{number of significant figures for IRR}

\item{CI_sf}{number of significant figures for 95\% CI}

\item{save_image_path}{file path to save table as image}

\item{save_docx_path}{file path to save table as docx}
}
\value{
flextable of AE summary by body system class and arm
}
\description{
\code{aetable} is used to plot a table of AE summary by body system class and arm using the following summary statistics:
\itemize{
\item \strong{Total number of participants at risk per arm}
\item \strong{Frequency} (number of participants with at least one event for each event)
\item \strong{Proportions} (number of participants with at least one event for each body system class relative to number of participants at risk)
\item \strong{Number of adverse events per participant}: presented as \strong{counts}, \strong{mean (SD)}
\item \strong{Total number of events} and \strong{incidence rates} (\strong{number of events} relative to \strong{total time in follow-up})
\item \strong{Treatment effect estimate (IRR)} and \strong{95\% confidence intervals}
}
}
\examples{
## ADAPT sample data
df2$aebodsys <- as.factor(df2$aebodsys)
aetable(df2, body_system_class="aebodsys", control="Placebo", intervention_levels=c("Intervention"), IRR=TRUE, variables = c("variable1", "variable2"))
}
