% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df2}
\alias{df2}
\title{Sample adverse event dataset with 2 treatment arms}
\format{
A data frame with 370 rows and 10 variables:
\describe{
\item{id}{Participant's id}
\item{arm}{Treatment arm: Placebo & Intervention}
\item{ae_pt}{Adverse event preferred term}
\item{aebodsys}{Body system class}
\item{severity}{Severity}
\item{date_rand}{Date of randomisation}
\item{date_ae}{Date of adverse event}
\item{last_visit}{Date of last visit}
\item{variable1}{Variable 1 to be included in treatment effect estimation}
\item{variable2}{Variable 2 to be included in treatment effect estimation}
}
}
\usage{
df2
}
\description{
Sample adverse event dataset with 2 treatment arms
}
\keyword{datasets}
